Address more corner cases that leave the async access state machine hung without sending a reply to attached waiters. Also cleaned up the type of the "pid" values used inside the locator.

+++ TAO/orbsvcs/ImplRepo_Service/ImR_Activator_i.cpp
@@ -357,7 +357,7 @@ ImR_Activator_i::kill_server (const char* name, CORBA::Long lastpid, CORBA::Shor
         ORBSVCS_DEBUG((LM_DEBUG,
                        "ImR Activator: Killing server <%C> "
                        "signal %d to pid %d, result = %d\n",
-                       name, signum, pid, result));
+                       name, signum, static_cast<int> (pid), result));
       if (!found && result == 0 && this->notify_imr_)
         {
           this->process_map_.bind (pid, name);
@@ -437,9 +437,9 @@ ImR_Activator_i::start_server(const char* name,
       if (debug_ > 1)
         ORBSVCS_DEBUG((LM_DEBUG,
                        "ImR Activator: Unique instance already running %d\n",
-                       pid));
+                       static_cast<int> (pid)));
       char reason[32];
-      ACE_OS::snprintf (reason,32,"pid:%d",pid);
+      ACE_OS::snprintf (reason,32,"pid:%d",static_cast<int> (pid));
       throw ImplementationRepository::CannotActivate(
         CORBA::string_dup (reason));
     }
@@ -500,7 +500,8 @@ ImR_Activator_i::start_server(const char* name,
       if (debug_ > 1)
         {
           ORBSVCS_DEBUG((LM_DEBUG,
-            "ImR Activator: register death handler for process %d\n", pid));
+                         "ImR Activator: register death handler for process %d\n",
+                         static_cast<int> (pid)));
         }
       this->process_map_.rebind (pid, name);
       if (unique)
@@ -527,7 +528,7 @@ ImR_Activator_i::start_server(const char* name,
     {
       ORBSVCS_DEBUG ((LM_DEBUG,
                       ACE_TEXT ("(%P|%t) ImR Activator: Successfully started <%C>, pid=%d\n"),
-                      name, pid));
+                      name, static_cast<int> (pid)));
     }
 }
 
@@ -553,7 +554,7 @@ ImR_Activator_i::handle_exit_i (pid_t pid)
           ORBSVCS_DEBUG ((LM_DEBUG,
                           ACE_TEXT ("(%P|%t) ImR Activator: Notifying ImR that ")
                           ACE_TEXT ("server[%d], <%C> has exited.\n"),
-                          pid, name.c_str()));
+                          static_cast<int> (pid), name.c_str()));
         }
       try
         {
